Refactor strain management in resistance module
- Consolidated imports in `tb.py` to streamline access to `StrainRegistry` and `StrainProfile`.
- Updated `__init__.py` to reflect the new location of `StrainProfile`, ensuring proper module exports.
- Removed `profile.py` as its functionality has been integrated into `strains.py`, enhancing code organization and reducing redundancy.

These changes improve the clarity and maintainability of the resistance module, facilitating better strain management in simulations.
